Move format and do_format out of BaseElement and remove unformatted
